Consolidate files. Remove dead code.
