--- src/hooks/use-measure-text.test.ts
@@ -1,76 +1,0 @@
-// TODO: Reinstate these tests.
-
-// // mock the measureText function
-// const mockMeasureText = jest.fn((text: string, fontSize: number) => {
-//   // assume every character is half the width of the font's height
-//   return { width: text.length * fontSize / 2 }
-// })
-
-// // mock the 2D canvas context
-// class MockCanvas2DContext {
-//   font = ""
-
-//   get fontSize() {
-//     const match = /(\d+)/.exec(this.font || "")
-//     const sizeStr = match?.[1]
-//     return sizeStr ? +sizeStr : 16
-//   }
-
-//   measureText(text: string) {
-//     return mockMeasureText(text, this.fontSize / 2)
-//   }
-// }
-
-// // mock document.createElement to return a "canvas" element that returns our mock 2D context
-// const origCreateElement = document.createElement
-// const mockCreateElement = jest.spyOn(document, "createElement").mockImplementation((tag: string, ...args: any[]) => (
-//   tag === "canvas"
-//     ? { getContext: () => new MockCanvas2DContext() } as any as HTMLCanvasElement
-//     : origCreateElement.call(document, tag, ...args)
-// ))
-
-// import { renderHook } from "@testing-library/react"
-// import { defaultFont } from "../components/constants"
-// import { measureText, useMeasureText } from "./use-measure-text"
-
-// describe("measureText", () => {
-//   const testText = "test text"
-//   const testText2 = "another string to test"
-//   const otherFont = "bold 20px serif"
-
-//   beforeEach(() => {
-//     mockMeasureText.mockClear()
-//   })
-
-//   afterAll(() => {
-//     mockCreateElement.mockRestore()
-//   })
-
-//   it("returns the same value for default font", () => {
-//     expect(measureText(testText)).toEqual(measureText(testText, defaultFont))
-//     // second call returns cached result
-//     expect(mockMeasureText).toHaveBeenCalledTimes(1)
-//   })
-//   it("returns different lengths for different strings", () => {
-//     expect(measureText(testText)).not.toEqual(measureText(testText2))
-//     // one result was cached by the previous test
-//     expect(mockMeasureText).toHaveBeenCalledTimes(1)
-//   })
-//   it("returns different lengths for different fonts", () => {
-//     expect(measureText(testText)).not.toEqual(measureText(testText, otherFont))
-//     // one result was cached by the previous test
-//     expect(mockMeasureText).toHaveBeenCalledTimes(1)
-//   })
-//   it("is consistent after multiple strings and fonts", () => {
-//     expect(measureText(testText)).toEqual(measureText(testText))
-//     // both results were cached by previous tests
-//     expect(mockMeasureText).toHaveBeenCalledTimes(0)
-//   })
-// })
-
-// describe("useMeasureText", () => {
-//   it("measures text when called as a hook", () => {
-//     const { result } = renderHook(() => useMeasureText())
-//     expect(result.current("word")).toBe(14)
-//   })
-// })
