--- src/plugins/xyplot/xyplot-defs.ts
@@ -1,3 +1,0 @@
-// TODO: Determine if these are redundant with content in xyplot-types.ts
-export const kGraphTileType = "CodapGraph";
-export const kGraphTileClass = "codap-graph";
