--- src/plugins/xyplot/xyplot-tile.test.tsx
@@ -1,48 +1,0 @@
-import { render } from "@testing-library/react";
-import React from "react";
-
-import { ITileApi } from "../../components/tiles/tile-api";
-import { TileModel } from "../../models/tiles/tile-model";
-import { defaultXYplotContent } from "./xyplot-content";
-import { XYplotToolComponent } from "./xyplot-tile";
-import "./xyplot-registration";
-
-describe("XYplotToolComponent", () => {
-  const content = defaultXYplotContent();
-  const model = TileModel.create({content});
-
-  const defaultProps = {
-    tileElt: null,
-    context: "",
-    docId: "",
-    documentContent: null,
-    isUserResizable: true,
-    onResizeRow: (e: React.DragEvent<HTMLDivElement>): void => {
-      throw new Error("Function not implemented.");
-    },
-    onSetCanAcceptDrop: (tileId?: string): void => {
-      throw new Error("Function not implemented.");
-    },
-    onRequestTilesOfType: (tileType: string): { id: string; title?: string | undefined; }[] => {
-      throw new Error("Function not implemented.");
-    },
-    onRequestUniqueTitle: (tileId: string): string | undefined => {
-      throw new Error("Function not implemented.");
-    },
-    onRequestRowHeight: (tileId: string, height?: number, deltaHeight?: number): void => {
-      throw new Error("Function not implemented.");
-    },
-    onRegisterTileApi: (tileApi: ITileApi, facet?: string): void => {
-      throw new Error("Function not implemented.");
-    },
-    onUnregisterTileApi: (facet?: string): void => {
-      throw new Error("Function not implemented.");
-    }
-  };
-
-  it("renders successfully", () => {
-    const {getByText} =
-      render(<XYplotToolComponent  {...defaultProps} {...{model}}></XYplotToolComponent>);
-    expect(getByText("XYplot Content Placeholder")).toBeInTheDocument();
-  });
-});
